Move test utilities to traits so they can be used in add-ons #2404
