+++ tests/includes/cf-test-case.php
@@ -9,8 +9,11 @@
  * @copyright 2016 Josh Pollock
  */
 class Caldera_Forms_Test_Case extends WP_UnitTestCase {
+	//Most utility functions are in traits so they can be included in other test suites.
+	use Caldera_Forms_Has_Mock_Form, Caldera_Forms_Imports_Form, Caldera_Forms_Has_Data;
 
 	const MOCK_FORM_ID = 'cf12345hiroy';
+
 
     /**
      * A form that isn't saved or on filter to use as a mock
@@ -24,7 +27,9 @@ class Caldera_Forms_Test_Case extends WP_UnitTestCase {
     /** @inheritdoc */
     public function setUp(){
 		$this->set_mock_form();
+		$this->mock_form_id = self::MOCK_FORM_ID;
 		//test that mock form is set properly
+		$this->assertSame( self::MOCK_FORM_ID, $this->mock_form_id );
 		$this->assertSame( self::MOCK_FORM_ID, $this->mock_form[ 'ID' ] );
 		$this->assertSame( 1, $this->mock_form[ 'pinned' ] );
 		parent::setUp();
@@ -80,301 +85,6 @@ class Caldera_Forms_Test_Case extends WP_UnitTestCase {
 			return $form;
 		}
 	}
-
-    /**
-     * Set mock_form property
-     *
-     * @since 1.5.7
-     */
-	private function set_mock_form(){
-        $this->mock_form = array(
-            'ID'                 => self::MOCK_FORM_ID,
-            'name'               => 'Another form',
-            'description'        => '',
-            'db_support'         => 1,
-            'pinned'             => 1,
-            'hide_form'          => 1,
-            'check_honey'        => 1,
-            'success'            => __( 'Form has been successfully submitted. Thank you.', 'caldera-forms' ),
-            'avatar_field'       => null,
-            'form_ajax'          => 1,
-            'custom_callback'    => '',
-            'layout_grid'        =>
-                array(
-                    'fields'    =>
-                        array(
-                            'fld_1724450' => '1:1',
-                            'fld_6125005' => '1:2',
-                            'fld_7269029' => '2:1',
-                            'fld_7896909' => '3:1',
-                        ),
-                    'structure' => '6:6|12#12',
-                ),
-            'fields'             =>
-                array(
-                    'fld_1724450' =>
-                        array(
-                            'ID'         => 'fld_1724450',
-                            'type'       => 'text',
-                            'label'      => 'Text',
-                            'slug'       => 'text',
-                            'conditions' =>
-                                array(
-                                    'type' => '',
-                                ),
-                            'caption'    => '',
-                            'config'     =>
-                                array(
-                                    'custom_class'  => '',
-                                    'placeholder'   => '',
-                                    'default'       => '',
-                                    'mask'          => '',
-                                    'type_override' => 'text',
-                                ),
-                        ),
-                    'fld_6125005' =>
-                        array(
-                            'ID'         => 'fld_6125005',
-                            'type'       => 'email',
-                            'label'      => 'Email',
-                            'slug'       => 'email',
-                            'conditions' =>
-                                array(
-                                    'type' => '',
-                                ),
-                            'required'   => 1,
-                            'caption'    => 'Make emails',
-                            'entry_list' => 1,
-                            'config'     =>
-                                array(
-                                    'custom_class' => '',
-                                    'placeholder'  => '',
-                                    'default'      => '',
-                                ),
-                        ),
-                    'fld_7269029' =>
-                        array(
-                            'ID'         => 'fld_7269029',
-                            'type'       => 'button',
-                            'label'      => 'Next Page',
-                            'slug'       => 'next_page',
-                            'conditions' =>
-                                array(
-                                    'type' => '',
-                                ),
-                            'caption'    => '',
-                            'config'     =>
-                                array(
-                                    'custom_class' => '',
-                                    'type'         => 'next',
-                                    'class'        => 'btn btn-default',
-                                    'target'       => '',
-                                ),
-                        ),
-                    'fld_7896909' =>
-                        array(
-                            'ID'         => 'fld_7896909',
-                            'type'       => 'button',
-                            'label'      => 'Submit',
-                            'slug'       => 'submit',
-                            'conditions' =>
-                                array(
-                                    'type' => '',
-                                ),
-                            'caption'    => '',
-                            'config'     =>
-                                array(
-                                    'custom_class' => '',
-                                    'type'         => 'submit',
-                                    'class'        => 'btn btn-default',
-                                    'target'       => '',
-                                ),
-                        ),
-                ),
-            'page_names'         =>
-                array(
-                    0 => 'Page 1',
-                    1 => 'Page 2',
-                ),
-            'conditional_groups' =>
-                array(
-                    'fields' =>
-                        array(),
-                ),
-            'settings'           =>
-                array(
-                    'responsive' =>
-                        array(
-                            'break_point' => 'sm',
-                        ),
-                ),
-            'mailer'             =>
-                array(
-                    'on_insert'     => 1,
-                    'sender_name'   => 'Caldera Forms Notification',
-                    'sender_email'  => 'admin@local.dev',
-                    'reply_to'      => '',
-                    'email_type'    => 'html',
-                    'recipients'    => '',
-                    'bcc_to'        => '',
-                    'email_subject' => 'Another form',
-                    'email_message' => '{summary}',
-                ),
-        );
-    }
-
-    /**
-     * Create a mock entry to test
-     *
-     * @since 1.4.0
-     */
-    protected function create_entry( array $form = null ){
-        if ( ! $form ) {
-            $form = $this->mock_form;
-        }
-        $data = array();
-        $i = 0;
-        foreach( $form[ 'fields' ] as $field_id => $field_config ){
-            if ( 1 == $i ) {
-                $data[ $field_id ] = $field_id . '_' . rand();
-            } else {
-                $data[ $field_id ] = array(
-                    rand(),
-                    5 => rand(), rand(), 'batman'
-                );
-            }
-            if( 0 == $i ){
-                $i = 1;
-            }else{
-                $i = 0;
-            }
-        }
-
-        $entry_id = Caldera_Forms_Save_Final::create_entry( $form, $data  );
-        return array(
-            'id' => $entry_id,
-            'field_data' => $data,
-            'form_id' => $form[ 'ID' ],
-        );
-    }
-
-    /**
-     * Import contact form without auto-responder
-     *
-     * @since 1.5.9
-     *
-     * @param bool $main_mailer Optional. If true, the default, contact form for main mailer is imported. If false, contact form for auti-responder is imported.
-     * @return string
-     */
-    protected function import_contact_form($main_mailer = true ){
-        if ($main_mailer) {
-            $file = $this->get_path_for_main_mailer_form_import();
-        } else {
-            $file = $this->get_path_for_auto_responder_contact_form_import();
-
-        }
-
-        return $this->import_form($file);
-    }
-
-
-
-    /**
-     * Cast array or object, like a form import, to array
-     *
-     * @since 1.5.9
-     *
-     * @param $array_or_object
-     * @return array
-     */
-    protected function recursive_cast_array( $array_or_object ){
-        $array_or_object = (array) $array_or_object;
-        foreach ( $array_or_object as $key => $value ){
-            if( is_array( $value ) || is_object( $value ) ){
-                $array_or_object[ $key ] = $this->recursive_cast_array( $value );
-            }
-
-        }
-        return $array_or_object;
-    }
-
-    /**
-     * Create submission data for mock submissions.
-     *
-     * Designed to be used to set $_POST for contact form tests or other mock submission requiring tests.
-     *
-     * @since 1.5.9
-     *
-     * @param null|string $form_id
-     * @param array $data
-     * @return array
-     */
-    protected function submission_data( $form_id = null, array $data = array() ){
-        if( ! $form_id ){
-            $form_id = self::MOCK_FORM_ID;
-        }
-
-        $nonce = Caldera_Forms_Render_Nonce::create_verify_nonce( $form_id );
-
-        $data = wp_parse_args( $data, array (
-            '_cf_verify' => $nonce,
-            '_wp_http_referer' => '/?page_id=4&preview=1&cf_preview=' . $form_id,
-            '_cf_frm_id' => $form_id,
-            '_cf_frm_ct' => '1',
-            'cfajax' => $form_id,
-            '_cf_cr_pst' => '4',
-            'email' => '',
-            'formId' => $form_id,
-            'instance' => '1',
-            'request' => site_url("/cf-api/$form_id"),
-            'postDisable' => '0',
-            'target' => '#caldera_notices_1',
-            'loadClass' => 'cf_processing',
-            'loadElement' => '_parent',
-            'hiderows' => 'true',
-            'action' => 'cf_process_ajax_submit',
-            'template' => "#cfajax_$form_id-tmpl",
-        ) );
-
-        return $data;
-    }
-
-    /**
-     * Import form by file path
-     *
-     * @since 1.5.9
-     *
-     * @param string $file Path to form config
-     * @return string
-     */
-    protected function import_form($file) {
-        $json = file_get_contents($file);
-        $config = $this->recursive_cast_array(json_decode($json));
-        $form_id = Caldera_Forms_Forms::import_form($config);
-        return $form_id;
-    }
-
-    /**
-     * Get file path for JSON export we import for contact form main mailer tests
-     *
-     * @since 1.6.0
-     *
-     * @return string
-     */
-    protected function get_path_for_main_mailer_form_import(){
-        return $file = dirname(__FILE__) . '/forms/contact-forms-no-auto-responder.json';
-    }
-
-    /**
-     * Get file path for JSON export we import for contact form auto-responder tests
-     *
-     * @since 1.6.0
-     *
-     * @return string
-     */
-    protected function get_path_for_auto_responder_contact_form_import(){
-        return dirname(__FILE__) . '/forms/contact-form-autoresponder.json';
-    }
 
     /**
      * Assert a value is numeric
